+++ src/Symfony/Component/Mime/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * [BC BREAK] Removed `NamedAddress` (`Address` now supports a name)
  * Added PHPUnit constraints
  * Added `AbstractPart::asDebugString()`
+ * Added `Address::fromString()`
 
 4.3.3
 -----